+++ package.js
@@ -23,9 +23,11 @@ Package.onUse(function(api) {
 Package.onTest(function(api) {
   configure(api);
   api.use('tinytest');
+  api.use('check');
   api.use('mongo');
   api.use('minimongo');
   api.use('http');
+  api.use('random');
   api.use('practicalmeteor:sinon');
   api.use('meteorhacks:fast-render');
   api.use('meteorhacks:inject-data');
@@ -56,8 +58,8 @@ function configure(api) {
   api.use('reactive-dict');
   api.use('reactive-var');
   api.use('ddp');
-
-  api.use('meteorhacks:fast-render@2.9.0', ['client', 'server']);
+  api.use('ejson');
+  api.use('meteorhacks:fast-render@2.10.0', ['client', 'server']);
   api.use('cosmos:browserify@0.5.0', 'client');
   api.use('meteorhacks:picker@1.0.3', 'server');
   api.use('meteorhacks:inject-data@1.3.0');